+++ webknossos/webknossos/dataset/layer/_downsampling_utils.py
@@ -344,7 +344,7 @@ def downsample_cube_job(
     try:
         source_bbox = source_view.normalized_bounding_box
         target_bbox = target_view.normalized_bounding_box
-        num_channels = target_view.info.bounding_box.size.c
+        num_channels = target_view.num_channels
         target_bbox_in_mag = target_bbox.in_mag(target_view.mag)
         shape = target_bbox_in_mag.size.to_tuple()
         shape_xyz = target_bbox_in_mag.size_xyz